# execute SQL updates as chunks in _set_spent function for tx_removals