@neurosis69 2022-05-24
I'll need to rewrite the script as the chosen `f"""` makes it way harder to do log/trace analysis.

This is how a sql log would look like with this PR
```sql
23:45:53.959293 SAVEPOINT s162960
23:45:53.959428
                            UPDATE OR FAIL coin_record INDEXED BY sqlite_autoindex_coin_record_1
                            SET spent_index=1556983
                            WHERE spent_index=0
                            AND coin_name IN (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)

23:45:53.959953 RELEASE s162960
```