@neurosis69 2022-05-24
> wow that is a major speedup! This must be one of the main bottlenecks of syncing, since there are probably ~1000 coins per transaction block during the dust storm periods.

Earle answered my question about the sqlite concerns in the last AMA mainly with the high amount of I/O which is done right now. I see a potential to reduce that, by optimizing SQLs and some parameters around.

For instance, I did full sync tests from height 0 to 1945145 on my i7-11700k recently and tested some of these changes.

branch | sync runtime | GB written
--- | --- | ---
main | 27h36m | 4351.769
test1 | 25h07m | 2443.256
test2 | 18h24m | 1009.895

I'd expect that on RPi the percentage would be even higher.
As I fully understand the reasoning behind the reduction of my test1, I am still not sure why test2 brings that much of an improvement. So I'm picking some of the kind of standalone changes and submitting them as PRs, like the last few I did.