@mariano54 2022-05-24
The tests are failing here, they need to pass. 
Make sure to throw the exception like the previous code did:
`            if ret.rowcount != len(coin_names):`

Those are some impressive improvements! what is test1 and what is test2?