@mariano54 2022-05-24
You can get rowcount from the Cursor object which is returned from execute I think.

We need to look at the improvements one by one to see which one has an impact.

The tests get trigerred every time you push a new commit. You can also close and re-open the PR to trigger the tests. 
The main thing is that you need to replicate the exact behavior of the previous code. For example, we use hex in the previous code, and not bytes for the coin names. Also, we throw a value error if we weren't able to set everything to spent. These things must be in the new PR.

Thanks for helping out btw, this is a very important change