@arvidn 2022-05-26
I'm interested in running benchmarks on this myself.