@arvidn 2022-05-26
running the `tools/test_full_sync.py`, I don't see any speed-up with this patch. I wonder why.

Running the stress test with 500 full blocks, I see:

| before | after |
| --- | --- |
| 154.56s | 155.57s |