@neurosis69 2022-05-26
On which hardware are you running your benchmarks? My timings in first post for this patch were taken from RPi.