@neurosis69 2022-05-26
I've started a few testsyncs (regular from peer) over the last few days and when I've sorted the data out I will share that data in my repo.

How can I start your mentioned benchmark? 
Just `python3 tools/test_full_sync.py`