@arvidn 2022-05-26
I'm running it on an AMD threadripper with SSD, so quite powerful compared to the RPi.

Unfortunately it's not so simple to run the benchmark. I'm using an artificial chain, created to max out all blocks. i.e. as a stress-test. You can create this chain by running:

```
python ./tools/generate_chain.py --length 500 --fill-rate 100
```

And when running `test_full_sync.py` you have to specify `--test-constants`, because this chain doesn't use mainnet constants, because that would be too expensive to create.