@neurosis69 2022-05-26
[chia-sync-data](https://github.com/neurosis69/chia-sync-data) is my repo where I'll upload the test data.

These uploads include the debug.log and automatically generated sar reports for every uploaded test.
Currently there are two testruns uploaded, one using the main branch and one with most of the optimizations I use (main_scenario0 branch).

The ansible repo is I use for these tests is [chia-sync-test](https://github.com/neurosis69/chia-sync-test).
Info about Scenarios/Testcases is also [there](https://github.com/neurosis69/chia-sync-test#scenarios-and-testcases)