@neurosis69 2022-05-26
For both testruns mentioned above, here is an additional grafana chart.

left side is using my branch with optimisations, right side is using main branch.

row 1: blocks synced over 5m
row 2: bytes written to db device
row 3: cpu utilization

![grafik](https://user-images.githubusercontent.com/83925572/170586237-e4959c9f-07dd-46c9-b7b8-38dd2ef89b1e.png)