@neurosis69 2022-05-26
> running the `tools/test_full_sync.py`, I don't see any speed-up with this patch. I wonder why.

I'm not sure what this test actually does? Is this a block validation test?
Because there is `await in_db.execute("pragma query_only")` for the only db connection.
When tracing the session only 1 SQL SELECT statements was issued.

I don't think that this is a valid test for my proposed change in _set_spent function, isn't it?