@neurosis69 2022-05-27
Finally I was able to run your benchmark on my i7-11700k with enabled sql logging.

before|after
---|---
148.28s|144.21s

sql.log output, taken from last updates issued on block height 479:

before:
```sql
13:38:03.322184 SAVEPOINT s2808
13:38:03.322345 UPDATE OR FAIL coin_record SET spent_index=? WHERE coin_name=? AND spent_index=0
...
13:38:03.348058 UPDATE OR FAIL coin_record SET spent_index=? WHERE coin_name=? AND spent_index=0
13:38:03.348209 RELEASE s2808
```
13:38:03.348209 - 13:38:03.322184 = 0,026025s

after
```sql
13:42:32.806548 SAVEPOINT s2808
13:42:32.807098 UPDATE OR FAIL coin_record INDEXED BY sqlite_autoindex_coin_record_1 SET spent_index=479 WHERE spent_index=0 AND coin_name IN (?,?,?,?,?,
?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?
...
,?,?,?,?,?,?,?,?,?,?)
13:42:32.814143 RELEASE s2808
```
13:42:32.814143 - 13:42:32.806548 = 0,007595s

I compared the main branch with the change `in _set_spent` function.
I wouldn't expect much IO related improvements by this PR, it is only about `UPDATE` execution time.

The IO related improvements would come from no-wal/synchronous and drop indexes during sync.
(although I'm currently testing the impact of letting wal on, as I had a few annoying db locking situations during my tests)