@neurosis69 2022-05-27
> I was expecting WAL would improve write performance, since updates are written sequentially to the WAL, and incorporated into the DB later, asynchronously.

Compared to other journal modes I'd agree. But compared to no journaling mode I'd expect it to be slower. At least you need additional checkpointing operation which slows down transactions. Probably different file format in WAL as well, but I don't know SQLite that good.

Here is hyperfine Benchmark result of 10 runs for above tests, slight improvement stays.
I'd expect a higher improvement percentage on RPi.

before:
```sh
Benchmark 1: python ./tools/test_full_sync.py run --test-constants /chia_temp1/test-chain/stress-test-blockchain-500-100-refs.sqlite
  Time (mean ± σ):     146.391 s ±  0.508 s    [User: 368.267 s, System: 147.680 s]
  Range (min … max):   145.375 s … 147.050 s    10 runs
```

after:
```sh
Benchmark 1: python ./tools/test_full_sync.py run --test-constants /chia_temp1/test-chain/stress-test-blockchain-500-100-refs.sqlite
  Time (mean ± σ):     143.878 s ±  0.536 s    [User: 364.189 s, System: 149.680 s]
  Range (min … max):   142.985 s … 144.855 s    10 runs
```